chore: rename baseCacheStore
