+++ packages/ponder/src/core/indexer/backfill.ts
@@ -1,7 +1,7 @@
 import { logger } from "@/common/logger";
 import { endBenchmark, startBenchmark } from "@/common/utils";
 import type { LogQueue } from "@/core/indexer/logQueue";
-import type { CacheStore } from "@/db/baseCacheStore";
+import type { CacheStore } from "@/db/cacheStore";
 import type { Source } from "@/sources/base";
 
 import { backfillSource } from "./backfillSource";
