+++ packages/ponder/src/core/indexer/historicalLogsRequestQueue.ts
@@ -2,7 +2,7 @@ import { BigNumber } from "ethers";
 import fastq from "fastq";
 
 import { logger } from "@/common/logger";
-import type { CacheStore } from "@/db/baseCacheStore";
+import type { CacheStore } from "@/db/cacheStore";
 import { parseLog } from "@/db/utils";
 import type { Source } from "@/sources/base";
