+++ packages/ponder/src/core/indexer/logQueue.ts
@@ -2,7 +2,7 @@ import { Contract } from "ethers";
 import fastq from "fastq";
 
 import { logger } from "@/common/logger";
-import type { CacheStore } from "@/db/baseCacheStore";
+import type { CacheStore } from "@/db/cacheStore";
 import type { Source } from "@/sources/base";
 import type { EventLog } from "@/types";
